Move {Open,Closed}ThreadOutput types
They are now only used as parameters to close and contest transaction
creation respectively.
