+++ hydra-tx/testlib/Test/Hydra/Tx/Gen.hs
@@ -23,7 +23,7 @@ import Hydra.Contract.HeadTokens (headPolicyId)
 import Hydra.Contract.Util (hydraHeadV1)
 import Hydra.Plutus (commitValidatorScript, initialValidatorScript)
 import Hydra.Tx (ScriptRegistry (..))
-import Hydra.Tx.CollectCom (OpenThreadOutput)
+import Hydra.Tx.Close (OpenThreadOutput)
 import Hydra.Tx.Crypto (Hash (..))
 import Hydra.Tx.Party (Party (..))
 import PlutusTx.Builtins (fromBuiltin)
